==> 1.Limpieza.ipynb <==
# Limpieza

import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import sidetable # si nos da un error este comando es por no tenerlo instalado
#!pip install holidays
import holidays
from datetime import date, datetime

pd.options.display.max_columns=None
# establecer el tamaño de nuestras figuras
plt.rcParams["figure.figsize"] = (10,8)

df = pd.read_csv('data/bikes.csv', index_col = 0)

df.head()

df.info()

df.isnull().sum()

df.duplicated().sum()

### Renombramos las columnas

Nuestro df contiene las siguientes columnas:
- registro = (instant), es el índice
- fecha = (dteday)
- estacion = (season) Hay que cambiarla, algunas estaciones no coinciden con la fecha.
- año = (year) También hay que cambiarla, está en 0  y 1 (2018,2019)
- mes = (month) 
- festivo = (holiday) 0: laborales, 1:festivos, creemos que tomará festivos.
- dia_semana = (weekday) 
- no_laboral = (workingday) Cuenta como 0 el laboral y el 1 el no laboral
- clima = ('weathersit')
- 'temperatura' =  (temp)
- 'sens_termica'=  ('atemp')
- humedad = (hum)
- viento = (windspeed)
- ocasionales = (casual)
- registrados = (registred)
- total = (cnt)


df.columns

dic_columnas = {'instant': 'registro', 'dteday': 'fecha', 'season': 'estacion', 'yr': 'año', 'mnth':'mes', 'holiday':'festivo', 'weekday':'dia_semana',
       'workingday':'no_laboral', 'weathersit':'clima', 'temp':'temperatura', 'atemp':'sens_termica', 'hum':'humedad', 'windspeed':'viento', 
       'casual':'ocasionales', 'registered':'registrados', 'cnt':'total'}

df.rename(columns = dic_columnas, inplace = True)

df.head(2)

df['festivo'].value_counts()

festivos = df[df['festivo'] == 1]
festivos.head(21)

#### Cambiamos la columna año, mes y día_semana

df['dia_semana'].unique()

Cambiamos el formato a la columna fecha

df['fecha_nueva'] = pd.to_datetime(df["fecha"], format="%d-%m-%Y")

df['festividad'] = pd.Series(df.fecha_nueva).apply(lambda x: holidays.CountryHoliday('US').get(x)).values

df['festividad_1'] = df['festividad'].astype('bool').astype('int')

df['festividad_1'].value_counts()

df.head(2)

Cambiamos las estaciones

Primero renombramos:

#df['año_bueno'] = df['fecha_nueva'].apply(datetime.date.today().year)

df['año_bueno'] = pd.DatetimeIndex(df['fecha_nueva']).year 

df['año_bueno'].value_counts()

df['año_bueno'].unique()

year = df['año_bueno'].unique()

for x in year:
    seasons = [('winter', (date(x,  1,  1),  date(x,  3, 20))),
            ('spring', (date(x,  3, 21),  date(x,  6, 20))),
            ('summer', (date(x,  6, 21),  date(x,  9, 22))),
            ('autumn', (date(x,  9, 23),  date(x, 12, 20))),
            ('winter', (date(x, 12, 21),  date(x, 12, 31)))]
print(seasons)


    

#print(get_season(date.today()))


def sacar_estacion(fecha):
    if fecha.month == 2 or fecha.month == 1:
        return "invierno"
    elif fecha.month == 3 and fecha.day <= 20:
        return "invierno"
    elif fecha.month == 12 and fecha.day >20:
        return "invierno"
    elif fecha.month == 4 or fecha.month == 5:
        return "primavera"
    elif fecha.month == 3 and fecha.day >20:
        return "primavera"
    elif fecha.month == 6 and fecha.day <=20:
        return "primavera"
    elif fecha.month == 7 or fecha.month == 8:
        return "verano"
    elif fecha.month == 6 and fecha.day > 20:
        return "verano"
    elif fecha.month == 9 and fecha.day < 23:
        return "verano"
    elif fecha.month == 10 or fecha.month == 11:
        return "otoño"
    elif fecha.month == 9 and fecha.day > 22:
        return "otoño"
    elif fecha.month == 12 and fecha.day <= 20:
        return "otoño"


    
    

df["estacion_correcta"] = df["fecha_nueva"].apply(sacar_estacion)

df.tail(15)

#Creamos un dataframe donde vamos a ir realizando los siguientes cambios.
df2=df.copy()
df2.head(3)

df2.drop(['estacion','festivo'],axis=1,inplace=True)

df2.head(1)

df2['mes_bueno'] = pd.DatetimeIndex(df2['fecha_nueva']).month
df2.head(1)

df2.drop(['fecha','año','mes'],axis=1,inplace=True)
df2.head(1)

df2['dia_semana_nuevo']=df2['fecha_nueva'].dt.weekday

df2.head(8)

df2.drop(['dia_semana'],axis=1,inplace=True)
df2.head(1)

df2['no_laboral'].value_counts()

#Creamos una funcion para crear una columna con los dias labroales.
def sacar_laboral(df):
    if (df['dia_semana_nuevo']<=4) and (df['festividad_1']== 0):
        return 1
    else:
        return 0

df2['no_laboral_nuevo']=df2.apply(sacar_laboral,axis=1)

df2.head(3)

df2['no_laboral_nuevo'].value_counts()
#Sí ha habido cambios en el numero de dias labroales detectados.

df3=df2.copy()
df3.head(3)

df3.drop(['no_laboral'],axis=1,inplace=True)
df3.head(3)

Limpiamos temperatura, sensación térmica, humedad y viento para dejarlas con 1 decimal.

df3['temperatura'] = df3['temperatura'].round(1)
df3['sens_termica'] = df3['sens_termica'].round(1)
df3['humedad'] = df3['humedad'].round(0)
df3['viento'] = df3['viento'].round(1)

df3.head(2)

df3.to_csv('data/bikes_limpio.csv')
