==> 4.7-codif_totales.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import sidetable # si nos da un error este comando es por no tenerlo instalado
from datetime import date, datetime
from scipy import stats
from sklearn.preprocessing import StandardScaler
pd.options.display.max_columns=None
# establecer el tamaño de nuestras figuras
plt.rcParams["figure.figsize"] = (10,8)

df = pd.read_csv("data/bikes_limpio.csv", index_col = 0)
df.head(2)

df.drop(['ocasionales', 'registrados'], axis = 1, inplace= True)
df.columns

#codificamos la columna año_bueno.
año_map={2018:0,2019:1}
df['año_map']=df['año_bueno'].map(año_map)
df.drop(['año_bueno'],axis=1,inplace=True)
df.sample(5)

#Codificamos la columna estacion_correcta.
estacion_map={'invierno':0,'primavera':1,'otoño':2,'verano':3}
df['estacion_map']=df['estacion_correcta'].map(estacion_map)
df.drop(['estacion_correcta'],axis=1,inplace=True)
df.sample(5)

df.to_csv('data/totales_encod.csv')
