==> 6.RF-registrados.ipynb <==
# Tratamiento de datos
# ------------------------------------------------------------------------------
import numpy as np
import pandas as pd

# Gráficos
# ------------------------------------------------------------------------------
import matplotlib.pyplot as plt
import seaborn as sns

# Modelado y evaluación
# ------------------------------------------------------------------------------
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn import tree
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.model_selection import GridSearchCV

# Barra de progreso de un proceso
# ------------------------------------------------------------------------------
from tqdm import tqdm

# Configuración warnings
# ------------------------------------------------------------------------------
import warnings
warnings.filterwarnings('once')

Primero codificados y estandarizados:

df=pd.read_csv('data/bikes_regis_estand_encod.csv',index_col=0)
df.head(3)

df.drop(['registro','fecha_nueva','festividad'], axis=1,inplace=True)
df.columns

X=df.drop(['registrados'],axis=1)
y=df['registrados']

# y dividir nuestros datos en train y test para poder evaluar la bondad de nuestro modelo

x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

param = {"max_depth": [5, 7, 8], 
         "max_features": [1,2,3], 
         "min_samples_split": [10, 75, 150], 
         "min_samples_leaf": [10, 75, 150]}

gs_rf = GridSearchCV(
            estimator=RandomForestRegressor(), # tipo de modelo que queremos hacer
            param_grid= param, # que hiperparámetros queremos que testee
            cv=10, # crossvalidation que aprendimos en la lección de regresión lineal intro. 
            verbose=-1, # para que no nos printee ningún mensaje en pantalla
            return_train_score = True, # para que nos devuelva el valor de las métricas de set de datos de entrenamiento
            scoring="neg_mean_squared_error") # la métrica que queremos que nos devuelva

# ajustamos el modelo que acabamos de definir en el GridSearch

gs_rf.fit(x_train, y_train)


bosque = gs_rf.best_estimator_
bosque

y_pred_test_rf = bosque.predict(x_test)
y_pred_train_rf = bosque.predict(x_train)

def metricas(y_test, y_train, y_test_pred, y_train_pred, tipo_modelo):
    
    
    resultados = {'MAE': [mean_absolute_error(y_test, y_test_pred), mean_absolute_error(y_train, y_train_pred)],
                'MSE': [mean_squared_error(y_test, y_test_pred), mean_squared_error(y_train, y_train_pred)],
                'RMSE': [np.sqrt(mean_squared_error(y_test, y_test_pred)), np.sqrt(mean_squared_error(y_train, y_train_pred))],
                'R2':  [r2_score(y_test, y_test_pred), r2_score(y_train, y_train_pred)],
                 "set": ["test", "train"]}
    df = pd.DataFrame(resultados)
    df["modelo"] = tipo_modelo
    return df

rf_encod_estand_results = metricas(y_test, y_train, y_pred_test_rf, y_pred_train_rf, "Random Forest registrados cod - estand")
rf_encod_estand_results

importancia_predictores = pd.DataFrame(
                            {'predictor': x_train.columns,
                             'importancia': bosque.feature_importances_}
                            )


# ordenamos de mayor a menor los resultados
importancia_predictores.sort_values(by=["importancia"], ascending=False, inplace = True)

# printeamos los resultados
print("Importancia de los predictores en el modelo")
print("-------------------------------------------")
importancia_predictores

## Ahora sin estandarizar:

df2 = pd.read_csv('data/bikes_regis_encod.csv', index_col=0)

df2.drop(['registro','fecha_nueva','festividad'], axis=1,inplace=True)
df2.columns

X2=df2.drop(['registrados'],axis=1)
y2=df2['registrados']

# Separamos en train y test
x2_train, x2_test, y2_train, y2_test = train_test_split(X2, y2, test_size=0.2, random_state=42)


param2 = {"max_depth": [5, 7, 8], 
         "max_features": [1,2,3], 
         "min_samples_split": [10, 75, 150], 
         "min_samples_leaf": [10, 75, 150]}

gs_rf2 = GridSearchCV(
            estimator=RandomForestRegressor(), # tipo de modelo que queremos hacer
            param_grid= param2, # que hiperparámetros queremos que testee
            cv=10, # crossvalidation que aprendimos en la lección de regresión lineal intro. 
            verbose=-1, # para que no nos printee ningún mensaje en pantalla
            return_train_score = True, # para que nos devuelva el valor de las métricas de set de datos de entrenamiento
            scoring="neg_mean_squared_error") # la métrica que queremos que nos devuelva

# ajustamos el modelo que acabamos de definir en el GridSearch
# 🚨 Esta celda puede tardar en ejecutarse

gs_rf2.fit(x2_train, y2_train)

bosque2 = gs_rf2.best_estimator_
bosque2

y2_pred_test_rf = bosque2.predict(x2_test)
y2_pred_train_rf = bosque2.predict(x2_train)

rf_resgistrados_encod = metricas(y2_test, y2_train, y2_pred_test_rf, y2_pred_train_rf, "Random Forest registrados codificados")
rf_resgistrados_encod 


# vamos a crearnos un dataframe, igual que hicimos en la clase anterior con la importancia de cada una de las variables incluidas en el modelo

importancia_predictores2 = pd.DataFrame(
                            {'predictor': x2_train.columns,
                             'importancia': bosque2.feature_importances_}
                            )


# ordenamos de mayor a menor los resultados
importancia_predictores2.sort_values(by=["importancia"], ascending=False, inplace = True)

# printeamos los resultados
print("Importancia de los predictores en el modelo")
print("-------------------------------------------")
importancia_predictores2
