==> 7_RF_totales_prediccion.ipynb <==
# Predicción empleando nuestro modelo predictivo

**Modelo**: RandomForest

**Variable respuesta**: 'Totales'

**Variables predictoras**: codificadas sin estandarizar.

# Tratamiento de datos
# ------------------------------------------------------------------------------
import numpy as np
import pandas as pd

# Gráficos
# ------------------------------------------------------------------------------
import matplotlib.pyplot as plt
import seaborn as sns

# Modelado y evaluación
# ------------------------------------------------------------------------------
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn import tree
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.model_selection import GridSearchCV

# Barra de progreso de un proceso
# ------------------------------------------------------------------------------
from tqdm import tqdm

# Configuración warnings
# ------------------------------------------------------------------------------
import warnings
warnings.filterwarnings('once')

# Poner descansos en código.
# ------------------------------------------------------------------------------
import time

# Librería para crear archivos pickle.
# ------------------------------------------------------------------------------
import pickle

df = pd.read_csv('data/totales_encod.csv', index_col = 0)
df.head(3)

df.drop(['registro','fecha_nueva','festividad','sens_termica'], axis=1,inplace=True)
df.columns

X=df.drop(['total'],axis=1)
y=df['total']

x_train, x_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Definir diccionario con los hiperparámetros a modificar
# min_samples_split: número mínimo de observaciones que debe tener un nodo para dividirse, por defecto [10, 50, 100]
# min_samples_leaf: número mínimo de observaciones que debe tener cada uno de los nodos hijos para que se produzca la division, por defecto [10, 50, 100]
param = {"max_depth": [16,17], "max_features": [1,2,3], "min_samples_split": [5, 75, 200], "min_samples_leaf": [5,75,200]}

gs_rf = GridSearchCV(estimator=RandomForestRegressor(), param_grid= param, cv=10, verbose=-1,
                        return_train_score = True, scoring="neg_mean_squared_error")

gs_rf.fit(x_train, y_train)
bosque = gs_rf.best_estimator_
bosque

def metricas(y_test, y_train, y_test_pred, y_train_pred, tipo_modelo):
    
    
    resultados = {'MAE': [mean_absolute_error(y_test, y_test_pred), mean_absolute_error(y_train, y_train_pred)],
                'MSE': [mean_squared_error(y_test, y_test_pred), mean_squared_error(y_train, y_train_pred)],
                'RMSE': [np.sqrt(mean_squared_error(y_test, y_test_pred)), np.sqrt(mean_squared_error(y_train, y_train_pred))],
                'R2':  [r2_score(y_test, y_test_pred), r2_score(y_train, y_train_pred)],
                 "set": ["test", "train"]}
    df = pd.DataFrame(resultados)
    df["modelo"] = tipo_modelo
    return df

y_pred_test_rf = bosque.predict(x_test)
y_pred_train_rf = bosque.predict(x_train)
df_totales_encod = metricas(y_test, y_train, y_pred_test_rf, y_pred_train_rf, "Random Forest totales codificadas")
df_totales_encod

## PROBLEMA: falta columna para año 2023

df.head(3)

######################## SOLUCIÓNAR PROBLEMA ###########################

#Codificamos de nuevo la columna año_bueno, ya que para poder hacer la prediccion neceitamos una columna con el año.
df2=df.copy()
df2['año_map_23']=0
df2.head(3)

#FUNCION CON ERROR QUE NO SE PUDO VOLVER A REPETIR ESTAS MÉTRICAS...
#def probar_modelo(dataframe,parametros):
#    X3=dataframe.drop(['total'],axis=1)
#    y3=dataframe['total']
#    time.sleep(3)
#    x3_train, x3_test, y3_train, y3_test = train_test_split(X3, y3, test_size=0.2, random_state=3)
#    time.sleep(3)
#    gs_rf3 = GridSearchCV(estimator=RandomForestRegressor(), param_grid= parametros, cv=10, verbose=-1,
#                        return_train_score = True, scoring="neg_mean_squared_error")
#    time.sleep(3)
#    gs_rf3.fit(x3_train, y3_train)
#    time.sleep(3)
#    bosque3 = gs_rf2.best_estimator_ #ERROR FUNCION, por eso salieron bien las metricas, es irreproducible ya...
#    time.sleep(3)
#    y3_pred_test_rf = bosque3.predict(x3_test)
#    y3_pred_train_rf = bosque3.predict(x3_train)
#    time.sleep(3)
#    df_totales_encod3 = metricas(y3_test, y3_train, y3_pred_test_rf, y3_pred_train_rf, "Random Forest totales codificadas")
#
#    return  bosque3,df_totales_encod3

#ERROR QUE NO SE PUDO VOLVER A REPETIR ESTAS MÉTRICAS...
#param5 = {"max_depth": [15,16,17], "max_features": [1,2,3], "min_samples_leaf": [2,3,5],"min_samples_split": [3, 5, 10]}
#bosque5,metricas5=probar_modelo(df2,param5)

#bosque5

#metricas5

#param4 = {"max_depth": [15], "max_features": [3], "min_samples_leaf": [2],"min_samples_split": [3]}
#bosque4,metricas4=probar_modelo(df2,param4)

### PRUEBA1

df2.head(3)

df3=df2.copy()

df3['año_map_18']=df3['año_map'].map({0:1,1:0})
df3['año_map_19']=df3['año_map']
df3.drop(['año_map'],axis=1,inplace=True)
df3.head(3)

def probar_modelo2(dataframe,parametros):
    X3=dataframe.drop(['total'],axis=1)
    y3=dataframe['total']
    time.sleep(3)
    x3_train, x3_test, y3_train, y3_test = train_test_split(X3, y3, test_size=0.2, random_state=9)
    time.sleep(3)
    gs_rf3 = GridSearchCV(estimator=RandomForestRegressor(), param_grid= parametros, cv=10, verbose=-1,
                        return_train_score = True, scoring="neg_mean_squared_error")
    time.sleep(3)
    gs_rf3.fit(x3_train, y3_train)
    time.sleep(3)
    bosque3 = gs_rf3.best_estimator_
    time.sleep(3)
    y3_pred_test_rf = bosque3.predict(x3_test)
    y3_pred_train_rf = bosque3.predict(x3_train)
    time.sleep(3)
    df_totales_encod3 = metricas(y3_test, y3_train, y3_pred_test_rf, y3_pred_train_rf, "Random Forest totales codificadas")

    return  bosque3,df_totales_encod3

param7 = {"max_depth": [15], "max_features": [3], "min_samples_leaf": [2],"min_samples_split": [1]}
bosque7,metricas7=probar_modelo2(df3,param7)

print(bosque7)
metricas7

df3.head(3)

X20=df3.drop(['total'],axis=1)
y20=df3['total']
time.sleep(3)
x20_train, x20_test, y20_train, y20_test = train_test_split(X20, y20, test_size=0.2, random_state=42)
time.sleep(3)
param20= {"max_depth": [15], "max_features": [3], "min_samples_leaf": [2],"min_samples_split": [1]}
gs_rf20 = GridSearchCV(estimator=RandomForestRegressor(), param_grid= param20, cv=10, verbose=-1,
                    return_train_score = True, scoring="neg_mean_squared_error")
time.sleep(3)
gs_rf20.fit(x20_train, y20_train)
time.sleep(3)
bosque20 = gs_rf20.best_estimator_
time.sleep(3)
y20_pred_test_rf = bosque20.predict(x20_test)
y20_pred_train_rf = bosque20.predict(x20_train)
time.sleep(3)
df_totales_encod20 = metricas(y20_test, y20_train, y20_pred_test_rf, y20_pred_train_rf, "Random Forest totales codificadas")

print(bosque20)
df_totales_encod20

df3.head(3)

### PRUEBA2 (NO USAR ESTA, MEJOR LA PRUEBA 1) (lo dejo puesto para que nos acordemos que ya se ha probado esto)

df2.head(3)

param10 = {"max_depth": [15], "max_features": [3], "min_samples_leaf": [2],"min_samples_split": [1]}
bosque10,metricas10=probar_modelo2(df2,param10)

print(bosque10)
metricas10

## Guardamos pickle del modelo

Guardaremos nuestro modelo en un archivo .pkl (pickle) para poder ejecutar el mismo modelo en otro ordenador pero sin perder el tiempo de ejecutar todo el modelo y correr el riesgo de perder esas métricas tan buenas.

bosque20

#Guardamos nuestro modelo en un pickle para no tener que correr el modelo cada vez que queramos
#  hacer una predicción.
with open('data/modelo_rf_totales.pkl', 'wb') as fp:
    pickle.dump(bosque20,fp)

## Continuació prediccion en 7_1_RF_totales_prediccion.ipynb
