==> 7_1_RF_totales_prediccion.ipynb <==
# Tratamiento de datos
# ------------------------------------------------------------------------------
import numpy as np
import pandas as pd
from datetime import date, datetime
import holidays

# Gráficos
# ------------------------------------------------------------------------------
import matplotlib.pyplot as plt
import seaborn as sns

# Modelado y evaluación
# ------------------------------------------------------------------------------
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn import tree
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.model_selection import GridSearchCV

# Barra de progreso de un proceso
# ------------------------------------------------------------------------------
from tqdm import tqdm

# Configuración warnings
# ------------------------------------------------------------------------------
import warnings
warnings.filterwarnings('once')

# Poner descansos en código.
# ------------------------------------------------------------------------------
import time

# Librería para crear archivos pickle.
# ------------------------------------------------------------------------------
import pickle
import os

## Abrimos en el modelo creado sin necesidad de volver a ejecutarlo.

Si ahora queremos ejecutar el mismo modelo en otro ordenador pero sin perder el tiempo de ejecutar todo el modelo y correr el riesgo de perder esas métricas tan buenas,deberemos hacer lo siguiente:

``` python
#Abrimos el pickle creado y aplicamos el modelo almacenado en él.
with open('data/modelo_rf_totales.pkl', 'rb') as f:
    forest=pickle.load(f)
```
``` python
#Comprobamos que el modelo se ha cargado adecuadamente.
forest
```
``` python
# Ya tendríamos el modelo cargado y solo tendríamos que pasarle el dataframe con los datos que queremos utilizar para predecir. IMPORTANTE: esos datos deben tener las mismas columnas que nuestros datos del X_train con el que creamos el modelo.
forest.predict(df_datos_prediccion)



#Abrimos el pickle creado y aplicamos el modelo almacenado en él.
with open('data/modelo_rf_totales.pkl', 'rb') as f:
    forest=pickle.load(f)

#Comprobamos que el modelo se ha cargado adecuadamente.
forest

Ya tendríamos el modelo cargado y solo tendríamos que pasarle el dataframe con los datos que queremos utilizar para predecir. IMPORTANTE: esos datos deben tener las mismas columnas que nuestros datos del X_train con el que creamos el modelo.
``` python
forest.predict(df_datos_prediccion)
```

## Prueba del modelo

prediccion={'clima': 2.0,
 'temperatura': 14.9,
 'humedad': 70.0,
 'viento': 16.7,
 'total': 801.0,
 'festividad_1': 0.0,
 'mes_bueno': 1.0,
 'dia_semana_nuevo': 1.0,
 'no_laboral_nuevo': 1.0,
 'estacion_map': 0.0,
 'año_map_23': 0.0,
 'año_map_18': 1.0,
 'año_map_19': 0.0}

df_pred=pd.DataFrame(prediccion,index=[0])
df_pred

x_pred=df_pred.drop(['total'],axis=1)

forest.predict(x_pred)

## Predicción

#### Estandarización
No tenemos que estandarizar nada ya que en nuestro modelo empleamos los datos sin estandarizar.

#### Encondings empleados en el modelo

- "año_bueno" fue codificada haciando la columna "año_map" donde:
``` python
        año_map={2018:0,2019:1}
        df['año_map']=df['año_bueno'].map(año_map)
        df.drop(['año_bueno'],axis=1,inplace=True)
```

- "estacion_correcta" fue codificada haciendo "estacion_map" donde:
``` python
        estacion_map={'invierno':0,'primavera':1,'otoño':2,'verano':3}
        df['estacion_map']=df['estacion_correcta'].map(estacion_map)
        df.drop(['estacion_correcta'],axis=1,inplace=True)
```

Por tanto para nuestra prediccion deberemos aplicar:
- **año**: en las predicciones tendremos el año 2023:

        · poner 1 en "año_map_23"
        · poner 0 en "año_map_18" y "año_map_19".
- **estación**: al recibir los datos con los que predecir, la estación será codificada de la misma manera que en el modelo:

``` python
        estacion_map={'invierno':0,'primavera':1,'otoño':2,'verano':3}
        df['estacion_map']=df['estacion_correcta'].map(estacion_map)
        df.drop(['estacion_correcta'],axis=1,inplace=True)
```

### Predicción con la API en 7_2_RF_totales_prediccion.ipynb
